infrastructure/docker/climatediscussionnexus-scrape/requirements.txt
beautifulsoup4
requests==2.25.0
pandas==1.1.4
newspaper3k==0.2.8
